% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cadaster.R
\name{download_cadaster}
\alias{download_cadaster}
\title{Download cadaster}
\usage{
download_cadaster(city_code, layer = "batiments")
}
\arguments{
\item{city_code}{A city code (string/number)}

\item{layer}{The layer we want to get (string)}
}
\value{
A file path to the downloaded json file
}
\description{
Downloads the json file containing the geo info according to the provided city_code and layer
}
\examples{
\dontrun{
download_cadaster("75001", "batiments")
}
}
