% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adress_data.R
\name{get_adress_data}
\alias{get_adress_data}
\title{gets detailed dataframe of the adresses that match a query adress}
\usage{
get_adress_data(query_adress)
}
\arguments{
\item{query_adress}{adresss to query}
}
\description{
gets detailed dataframe of the adresses that match a query adress
}
\examples{
\dontrun{
get_adress_data("12, rue de Lota 75116 Paris")
}
}
