% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area.R
\name{get_area}
\alias{get_area}
\title{Get area}
\usage{
get_area(sf_dataframe)
}
\arguments{
\item{sf_dataframe}{A SF dataframe with geometries}
}
\value{
A "units" object
}
\description{
Computes the area of a list of Single Feature geometries
}
\examples{
\dontrun{
get_area(sf_dataframe)
}
}
