% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_areas.R
\name{get_areas}
\alias{get_areas}
\title{get_areas}
\usage{
get_areas(polygon_df)
}
\arguments{
\item{polygon_df}{your SpatialPolygonsDataFrame}
}
\value{
returns the area of the polygon of a given address
}
\description{
get_areas
}
\examples{
\dontrun{
get_areas(polygon_df)
}
}
