% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bano_matches.R
\name{get_bano_matches}
\alias{get_bano_matches}
\title{Get BANO matches}
\usage{
get_bano_matches(address)
}
\arguments{
\item{address}{Address to query}
}
\value{
A dataframe containing BANO search results
}
\description{
Retrieves all addresses (+ useful info) that match the address query in the BANO database
}
\examples{
\dontrun{
get_bano_matches("12, rue de Lota 75116 Paris")
}
}
