% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cadaster_sp.R
\name{get_cadaster_sp}
\alias{get_cadaster_sp}
\title{returns sp::PolygonesDataFrame with all the infos on the cadaster type of a "commune"}
\usage{
get_cadaster_sp(communal_code, cadaster_type = "batiments")
}
\arguments{
\item{communal_code}{code of the desired city}

\item{cadaster_type}{type of cadaster to extract}
}
\description{
returns sp::PolygonesDataFrame with all the infos on the cadaster type of a "commune"
}
\examples{
\dontrun{
get_cadaster_sp(75116, cadaster_type = "parcelles")
}
}
