% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feuille_id.R
\name{get_feuille_id}
\alias{get_feuille_id}
\title{get_feuille_id
Fonction qui donne la liste des numéros de feuilles pour un SpatialPolygonDataframe construit avec le cadastre}
\usage{
get_feuille_id(polygon, polygon_df)
}
\arguments{
\item{polygon}{: SpatialPolygoneDataFrame, output de get_cadaster_sp (cadaster_type = "parcelles")}

\item{polygon_df}{: polygone spécifique, output de la fonction select_polygon}
}
\description{
get_feuille_id
Fonction qui donne la liste des numéros de feuilles pour un SpatialPolygonDataframe construit avec le cadastre
}
