% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearest_polygon.R
\name{get_nearest_polygon}
\alias{get_nearest_polygon}
\title{Get nearest polygon}
\usage{
get_nearest_polygon(sf_dataframe, long, lat)
}
\arguments{
\item{sf_dataframe}{dataframe containing polygons to choose from}

\item{long}{longitude of the point}

\item{lat}{latitude of the point}
}
\value{
sf_dataframe with the closest polygon
}
\description{
Get nearest polygon
}
\examples{
\dontrun{
get_nearest_polygon(town, adress_data$longitude, adress_data$latitude)
}
}
