% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parcelle_number.R
\name{get_parcelle_number}
\alias{get_parcelle_number}
\title{Get parcelle number}
\usage{
get_parcelle_number(poly)
}
\arguments{
\item{poly}{: sf_data frame with only one polygone}
}
\value{
The polygon number
}
\description{
Fonction qui donne le numero de la parcelle
}
\examples{
\dontrun{
get_feuille_id(polygon)
}
}
