% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parent_polygon.R
\name{get_parent_polygon}
\alias{get_parent_polygon}
\title{Get parent polygon}
\usage{
get_parent_polygon(polygon, sf_dataframe)
}
\arguments{
\item{polygon}{A sfg object (single polygon)}

\item{sf_dataframe}{A SF dataframe}
}
\value{
A 1*n SF dataframe containing the matched line
}
\description{
Gets the shape of the sf_dataframe that intersects the given polygon
}
\examples{
\dontrun{
feuille <- get_parent_polygon(building, parcelles)
}
}
