% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_section.R
\name{get_section}
\alias{get_section}
\title{get_section}
\usage{
get_section(polygon)
}
\arguments{
\item{polygon}{: SpatialPolygoneDataFrame de l'adresse (cadaster_type = "sections" quand on get_cadaster_sp)}
}
\description{
get_section
}
