% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_sf.R
\name{json_to_sf}
\alias{json_to_sf}
\title{JSON to SF}
\usage{
json_to_sf(json_file)
}
\arguments{
\item{json_file}{A path to the JSON file to parse}
}
\value{
A SF dataframe
}
\description{
Parses the provided GeoJSON to a Simple Features dataframe
}
\examples{
\dontrun{
df <- json_to_sf("../../cadaster-75001.json")
}
}
