% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_from_list.R
\name{polygon_from_list}
\alias{polygon_from_list}
\title{Polygon from list}
\usage{
polygon_from_list(list_of_points)
}
\arguments{
\item{list_of_points}{List of point vectors}
}
\value{
a sfc
}
\description{
Creates polygon from coordinate list
}
\examples{
\dontrun{
ls <- list(c(long, lat), c(long+ 0.001, lat), c(long+0.001, lat +0.0005))
poly <- polygon_from_list(ls)
}
}
