% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adress_data.R
\name{get_adress_data}
\alias{get_adress_data}
\title{Get address data}
\usage{
get_adress_data(query_adress)
}
\arguments{
\item{query_adress}{address to query}
}
\value{
A dataframe
}
\description{
Retrieves add addresses (+ useful info) that match the address query
}
\examples{
\dontrun{
get_adress_data("12, rue de Lota 75116 Paris")
}
}
