% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cadaster_sp.R
\name{get_cadaster_sp}
\alias{get_cadaster_sp}
\title{Get cadaster SP}
\usage{
get_cadaster_sp(city_code, cadaster_type = "batiments")
}
\arguments{
\item{cadaster_type}{type of cadaster to extract}

\item{citycode}{code of the desired city}
}
\value{
sp::SpatialPolygonsDataFrame
}
\description{
Downloads the cadaster GeoJSON file associated with a city code and a data type (parcelles, etc.)
Returns a SP dataframe containing the geographic data
}
\examples{
\dontrun{
get_cadaster_sp(75001, cadaster_type = "parcelles")
}
}
