% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cadaster_sp.R
\name{get_cadaster_sp}
\alias{get_cadaster_sp}
\title{Get cadaster SP (DEPRECATED)}
\usage{
get_cadaster_sp(city_code, layer = "batiments")
}
\arguments{
\item{city_code}{A city code}

\item{layer}{A layer}
}
\value{
A SP dataframe
}
\description{
Wrapper around download_cadaster and json_to_sf functions
}
\examples{
\dontrun{
get_cadaster_sp(75001, "batiments")
}
}
