% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feuille_id.R
\name{get_feuille_id}
\alias{get_feuille_id}
\title{get_feuille_id
Fonction qui donne le numero de feuille <- pas encore commit ce commentaire !}
\usage{
get_feuille_id(polygon)
}
\arguments{
\item{polygon}{: SpatialPolygoneDataFrame, output de get_cadaster_sp (cadaster_type = "feuilles")}
}
\value{
The "feuille" number
}
\description{
get_feuille_id
Fonction qui donne le numero de feuille <- pas encore commit ce commentaire !
}
\examples{
\dontrun{
get_feuille_id(polygon)
}
}
