% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Get ID}
\usage{
get_id(poly)
}
\arguments{
\item{poly}{sf_data frame with only one polygone}
}
\value{
The polygon id number
}
\description{
Fonction qui donne le numero de feuille
}
\examples{
\dontrun{
get_feuille_id(polygon)
}
}
