% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearest_polygon.R
\name{get_nearest_polygon}
\alias{get_nearest_polygon}
\title{Get nearest polygon}
\usage{
get_nearest_polygon(sf_dataframe, long, lat)
}
\arguments{
\item{sf_dataframe}{A dataframe containing polygons to choose from}

\item{long}{The longitude of the point}

\item{lat}{The latitude of the point}
}
\value{
A 1*N SF dataframe containing the closest polygon
}
\description{
gets the nearest polygon from a long lat point in a given SF dataframe
}
\examples{
\dontrun{
polygon <- get_nearest_polygon(town, adress_data$longitude, adress_data$latitude)
}
}
