% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_section.R
\name{get_section}
\alias{get_section}
\title{Get section}
\usage{
get_section(polygon)
}
\arguments{
\item{polygon}{: SpatialPolygoneDataFrame de l'adresse (cadaster_type = "sections" quand on get_cadaster_sp)}
}
\value{
section of the town of a polygon
}
\description{
Get section
}
\examples{
\dontrun{
sections_df <- cadastertools::get_cadaster_sp(adress_data$citycode, layer = "sections")
section <- get_nearest_polygon(section_df, adress_data$longitude, adress_data$latitude)
section_name <- get_section(section)
}
}
