% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_from_list.R
\name{polygon_from_list}
\alias{polygon_from_list}
\title{polygon_from_list : create polygon from coordinate list}
\usage{
polygon_from_list(point_list)
}
\arguments{
\item{point_list}{coordinate list}
}
\value{
sfc polygon
}
\description{
polygon_from_list : create polygon from coordinate list
}
\examples{
\dontrun{
ls <- list(c(long, lat), c(long+ 0.001, lat), c(long+0.001, lat +0.0005))
poly <- polygon_from_list(ls)
}
}
