% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_polygone.R
\name{select_polygone}
\alias{select_polygone}
\title{selects the polygon whoch contains the lattitude and the lattitude specified in a SpatialPolygoneDataFrame}
\usage{
select_polygone(sp_data, long, lat)
}
\arguments{
\item{sp_data}{the SpatialPolygoneDataFrame}

\item{long}{the longitude}

\item{lat}{the latitude}
}
\description{
selects the polygon whoch contains the lattitude and the lattitude specified in a SpatialPolygoneDataFrame
}
\examples{
\dontrun{
select_polygone(12.05, 14.215, my_data)
}
}
